MA-6991 fix Screen flash on IMX6 platform.
The root cause is BufferQueue slot mechanism is changed to
get free buffer. When BufferQueueConsumer consume buffers
fast enough to provide free buffer to BufferQueueProducer,
two buffers are sufficient to meet consumer and producer.
But display requires three buffers to avoid screen tearing.
So, allocate enough buffers to meet display requirement.

This buffer queue mechanism is changed on android 7.0.
So, redo this commit to use 3 framebuffer.

Change-Id: I9c52e276139677bcd36136642259d32f46f6c299
Signed-off-by: Xiaowen Liu <xiaowen.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/991
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>